Move weather service to Services
