+++ src/MeteoWeatherAPI/Program.cs
@@ -1,5 +1,4 @@
 using Application.Cache;
-using Application.Service;
 using Application.Validators;
 using DataAccess.Context;
 using DataAccess.Services;
@@ -7,6 +6,7 @@ using Domain.Services;
 using FluentValidation;
 using FluentValidation.AspNetCore;
 using MeteoWeatherAPI.CustomActionFilter;
+using MeteoWeatherAPI.Services;
 using MongoDB.Driver;
 
 var builder = WebApplication.CreateBuilder(args);
